37775 - 64 bit /*long*/ markup

+++ bundles/org.eclipse.swt/Eclipse SWT Drag and Drop/gtk/org/eclipse/swt/dnd/Transfer.java
@@ -127,6 +127,6 @@ abstract protected Object nativeToJava(TransferData transferData);
 public static int registerType(String formatName){
 	if (formatName == null) return OS.GDK_NONE;
 	byte[] buffer = Converter.wcsToMbcs(null, formatName, true);
-	return OS.gdk_atom_intern(buffer, false);
+	return (int)/*64*/OS.gdk_atom_intern(buffer, false);
 }
 }
